fix: remove Sentry, consolidate env files, update Lighthouse metrics
- Remove Sentry stub (src/lib/sentry.ts) and unused error-tracking.ts
- Create clean src/lib/error-tracking.ts with captureException API
- Update 4 error boundary imports from @/lib/sentry to @/lib/error-tracking
- Remove all Sentry references from .env.example, IMPLEMENTATION_PLAN, CSP route
- Remove legacy ADMIN_PASSWORD/ADMIN_USERNAME from .env.example
- Remove completed GitHub Branch Protection checklist
- Update IMPLEMENTATION_PLAN with Feb 25 Lighthouse results (43/100 perf)
- Mark ADMIN_TOTP_SECRET_SALT as configured in IMPLEMENTATION_PLAN
- Add ADMIN_TOTP_SECRET_SALT to .env.example
