+++ .env.example
@@ -7,14 +7,6 @@
 # Authentication (REQUIRED for production)
 # =============================================================================
 
-# --- Legacy Basic Auth (DEPRECATED - will be removed) ---
-# Admin password - MUST be strong (12+ chars, upper/lower/number/special)
-# Generate with: npm run generate:secrets
-# ADMIN_PASSWORD=
-# ADMIN_USERNAME=admin
-
-# --- New Database-Backed Authentication ---
-
 # Dev-only: bypass admin authentication (NEVER use in production)
 # Set to "true" in .env.local to skip login during development
 # ADMIN_DEV_BYPASS=true
@@ -33,6 +25,10 @@ NEXTAUTH_URL=http://localhost:3000
 # MFA encryption key for TOTP secrets (64-character hex string = 32 bytes)
 # Generate with: openssl rand -hex 32
 MFA_ENCRYPTION_KEY=
+
+# TOTP secret salt for admin authentication (32-character hex string = 16 bytes)
+# Generate with: openssl rand -hex 16
+ADMIN_TOTP_SECRET_SALT=
 
 # =============================================================================
 # Rate Limiting (REQUIRED for production)
@@ -61,23 +57,6 @@ UPSTASH_REDIS_REST_TOKEN=
 
 # CSP Violation Reporting (optional)
 # CSP_REPORT_URI=https://your-logging-service.com/csp-report
-
-# =============================================================================
-# Error Tracking (Sentry)
-# =============================================================================
-
-# Sentry DSN for error tracking (optional but recommended for production)
-# Get from: https://sentry.io → Settings → Projects → [Project] → Client Keys
-# Format: https://<key>@o<org-id>.ingest.sentry.io/<project-id>
-# NEXT_PUBLIC_SENTRY_DSN=
-
-# Sentry auth token for source maps upload (optional)
-# Get from: https://sentry.io → Settings → Auth Tokens
-# SENTRY_AUTH_TOKEN=
-
-# Sentry organization and project slugs (for source maps)
-# SENTRY_ORG=your-org
-# SENTRY_PROJECT=costa-rica-tree-atlas
 
 # =============================================================================
 # Public Variables (exposed to browser)
